move optical property scripts to subdir and update import paths
