+++ tests/test_snicar.py
@@ -29,7 +29,7 @@ import numpy as np
 import pytest
 from biosnicar.rt_solvers.adding_doubling_solver import adding_doubling_solver
 from biosnicar.classes import Impurity
-from biosnicar.column_OPs import get_layer_OPs, mix_in_impurities
+from biosnicar.optical_properties.column_OPs import get_layer_OPs, mix_in_impurities
 from biosnicar.setup_snicar import setup_snicar
 from biosnicar.rt_solvers.toon_rt_solver import toon_solver
